requirements.txt
Flask==2.2.5
paho-mqtt==1.6.1